feat(quota): 迁移额度管理插件
